requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
opencv-python
numpy
pillow
